clean the implementation of tinykv project
