+++ kv/server/raw_api.go
@@ -14,121 +14,26 @@ import (
 // RawGet return the corresponding Get response based on RawGetRequest's CF and Key fields
 func (server *Server) RawGet(_ context.Context, req *kvrpcpb.RawGetRequest) (*kvrpcpb.RawGetResponse, error) {
 	// Your Code Here (1).
-	resp := &kvrpcpb.RawGetResponse{}
-	reader, err := server.storage.Reader(req.Context)
-	if err != nil {
-		if regionErr, ok := err.(*raft_storage.RegionError); ok {
-			resp.RegionError = regionErr.RequestErr
-			return resp, nil
-		}
-		return nil, err
-	}
-	defer reader.Close()
-	val, err := reader.GetCF(req.Cf, req.Key)
-	if err != nil {
-		if regionErr, ok := err.(*raft_storage.RegionError); ok {
-			resp.RegionError = regionErr.RequestErr
-			return resp, nil
-		}
-		return nil, err
-	}
-	resp.Value = val
-	if val == nil {
-		resp.NotFound = true
-	}
-	return resp, nil
+	return nil, nil
 }
 
 // RawPut puts the target data into storage and returns the corresponding response
 func (server *Server) RawPut(_ context.Context, req *kvrpcpb.RawPutRequest) (*kvrpcpb.RawPutResponse, error) {
 	// Your Code Here (1).
 	// Hint: Consider using Storage.Modify to store data to be modified
-	resp := &kvrpcpb.RawPutResponse{}
-	batch := []storage.Modify{
-		{
-			Data: storage.Put{
-				Cf:    req.Cf,
-				Key:   req.Key,
-				Value: req.Value,
-			},
-		}}
-	err := server.storage.Write(req.Context, batch)
-	if err != nil {
-		if regionErr, ok := err.(*raft_storage.RegionError); ok {
-			resp.RegionError = regionErr.RequestErr
-			return resp, nil
-		}
-		return nil, err
-	}
-	return resp, nil
+	return nil, nil
 }
 
 // RawDelete delete the target data from storage and returns the corresponding response
 func (server *Server) RawDelete(_ context.Context, req *kvrpcpb.RawDeleteRequest) (*kvrpcpb.RawDeleteResponse, error) {
 	// Your Code Here (1).
 	// Hint: Consider using Storage.Modify to store data to be deleted
-	resp := &kvrpcpb.RawDeleteResponse{}
-	batch := []storage.Modify{
-		{
-			Data: storage.Delete{
-				Cf:  req.Cf,
-				Key: req.Key,
-			},
-		}}
-	err := server.storage.Write(req.Context, batch)
-	if err != nil {
-		if regionErr, ok := err.(*raft_storage.RegionError); ok {
-			resp.RegionError = regionErr.RequestErr
-			return resp, nil
-		}
-		return nil, err
-	}
-	return resp, nil
+	return nil, nil
 }
 
 // RawScan scan the data starting from the start key up to limit. and return the corresponding result
 func (server *Server) RawScan(_ context.Context, req *kvrpcpb.RawScanRequest) (*kvrpcpb.RawScanResponse, error) {
 	// Your Code Here (1).
 	// Hint: Consider using reader.IterCF
-	resp := &kvrpcpb.RawScanResponse{}
-	if req.Limit == 0 {
-		return resp, nil
-	}
-	reader, err := server.storage.Reader(req.Context)
-	if err != nil {
-		if regionErr, ok := err.(*raft_storage.RegionError); ok {
-			resp.RegionError = regionErr.RequestErr
-			return resp, nil
-		}
-		return nil, err
-	}
-	defer reader.Close()
-	iter := reader.IterCF(req.Cf)
-	if err != nil {
-		if regionErr, ok := err.(*raft_storage.RegionError); ok {
-			resp.RegionError = regionErr.RequestErr
-			return resp, nil
-		}
-		return nil, err
-	}
-	defer iter.Close()
-	var pairs []*kvrpcpb.KvPair
-	n := req.Limit
-	for iter.Seek(req.StartKey); iter.Valid(); iter.Next() {
-		item := iter.Item()
-		val, err := item.ValueCopy(nil)
-		if err != nil {
-			return nil, err
-		}
-		pairs = append(pairs, &kvrpcpb.KvPair{
-			Key:   item.KeyCopy(nil),
-			Value: val,
-		})
-		n--
-		if n == 0 {
-			break
-		}
-	}
-	resp.Kvs = pairs
-	return resp, nil
+	return nil, nil
 }
