+++ kv/storage/standalone_storage_ldb/standalone_storage.go
@@ -1,0 +1,42 @@
+package standalone_storage_ldb
+
+import (
+	"fmt"
+
+	"github.com/jmhodges/levigo"
+	"github.com/pingcap-incubator/tinykv/kv/config"
+	"github.com/pingcap-incubator/tinykv/kv/storage"
+	"github.com/pingcap-incubator/tinykv/kv/util/engine_util"
+	"github.com/pingcap-incubator/tinykv/proto/pkg/kvrpcpb"
+)
+
+// StandAloneStorage is an implementation of `Storage` for a single-node TinyKV instance. It does not
+// communicate with other nodes and all data is stored locally.
+type StandAloneStorage struct {
+	// Your Data Here (1).
+}
+
+func NewStandAloneStorage(conf *config.Config) *StandAloneStorage {
+	// Your Code Here (1).
+	return nil
+}
+
+func (s *StandAloneStorage) Start() error {
+	// Your Code Here (1).
+	return nil
+}
+
+func (s *StandAloneStorage) Stop() error {
+	// Your Code Here (1).
+	return nil
+}
+
+func (s *StandAloneStorage) Reader(ctx *kvrpcpb.Context) (storage.StorageReader, error) {
+	// Your Code Here (1).
+	return nil, nil
+}
+
+func (s *StandAloneStorage) Write(ctx *kvrpcpb.Context, batch []storage.Modify) error {
+	// Your Code Here (1).
+	return nil
+}
